Test Feemarket parameter change
Added test for the feemarket param change.

This test is currently failing during the proposal submission stage
`submitGovProsposal`.
The proposal that is being sent is generated using the draft-proposal
feature, hence the formatting should be correct.
The problem seems to be that the Hex of the transaction does not appear
in the chain after the submission.